fix a few bugs
