+++ src/oop.my.graphs.src.gresource.xml
@@ -5,6 +5,9 @@
     <file>window.js</file>
     <file>sidebar.js</file>
     <file>area.js</file>
+    <file>style.js</file>
+    <file>graph-color-button.js</file>
+    <file>primitives.js</file>
     <file>mirror.js</file>
   </gresource>
 </gresources>
